requirements/tests.txt
factory_boy==2.11.1
pytest==4.2.1
pytest-django==3.4.8
pytest-pythonpath==0.7.3
tox==3.7.0
flake8==3.7.7
codecov==2.0.15
pytest-cov==2.6.1
